backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "My FastAPI application"
authors = ["Nelson ALIKOU <nelsonalikou@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
psycopg2-binary = "^2.9.9"
SQLAlchemy = "^2.0.29"
pydantic = {version = "^2.7.2", extras = ["email"]}
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
bcrypt = "^4.3.0"

[tool.poetry.dev-dependencies]
# Add any development dependencies here, e.g.:
# pytest = "^7.4.4"
# mypy = "^1.10.0"
# black = "^24.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
